refactor: Change folder name
